Handle error cases when logs are not created
When the glob cannot be expanded, due to no files matching the path,
the glob value itself is the output.

Example:
for x in "foo/*"; do
    echo $x
done

The above will echo the contents of "foo" if they exist, otherwise it
will output "foo/*".  This was causing the script to attempt copying
non-existent files in logs when `cvd load` appears to be failing early.

We could also have set `nullglob` to turn the for loops into no-ops when
there are no files expanded by the pattern. I was not sure howthe option
would affect other parts of the script and went with this local fix.

+++ e2etests/cvd/cvd_load_boot_test.sh
@@ -36,18 +36,28 @@ function collect_logs_and_cleanup() {
     cp "${ENV_FILE}" "${TEST_UNDECLARED_OUTPUTS_DIR}/environment.json"
     cp "${CMD_OUT}" "${TEST_UNDECLARED_OUTPUTS_DIR}"
     cp "${CMD_ERR}" "${TEST_UNDECLARED_OUTPUTS_DIR}"
+
     # TODO(b/324650975): cvd doesn't print very useful information yet so file locations must be extracted this way
     home_dir="$(grep -o -E -m1 HOME="/tmp/cvd/[0-9a-zA-Z/_]+" "${CMD_ERR}" | cut -d= -f2)"
     artifacts_dir="$(grep -o -E '\-\-target_directory=/tmp/cvd/[0-9a-zA-Z/_]+' "${CMD_ERR}" | cut -d= -f2 | grep -o -E '[^"]*')"
 
     cp "${artifacts_dir}/fetch.log" "${TEST_UNDECLARED_OUTPUTS_DIR}"
+
     for instance_dir in "${home_dir}"/cuttlefish/instances/*; do
-      instance="$(basename "${instance_dir}")"
-      for log_file in "${instance_dir}"/logs/*; do
-        base_name="$(basename "${log_file}")"
-        cp "${log_file}" "${TEST_UNDECLARED_OUTPUTS_DIR}/${instance}_${base_name}"
-      done
-      cp "${instance_dir}"/cuttlefish_config.json "${TEST_UNDECLARED_OUTPUTS_DIR}/${instance}_cuttlefish_config.json"
+      if [[ -d "${instance_dir}" ]]; then
+        instance="$(basename "${instance_dir}")"
+        for log_file in "${instance_dir}"/logs/*; do
+          if [[ -f "${log_file}" ]]; then
+            base_name="$(basename "${log_file}")"
+            cp "${log_file}" "${TEST_UNDECLARED_OUTPUTS_DIR}/${instance}_${base_name}"
+          fi
+        done
+        if [[ -f "${instance_dir}/cuttlefish_config.json" ]]; then
+          cp "${instance_dir}"/cuttlefish_config.json "${TEST_UNDECLARED_OUTPUTS_DIR}/${instance}_cuttlefish_config.json"
+        else
+          echo "No cuttlefish_config.json found for instance ${instance}"
+        fi
+      fi
     done
   fi
